backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic[email]==2.10.3
python-multipart==0.0.19
fastapi-mail==1.4.1
redis==4.6.0
fastapi-cache2[redis]==0.2.1
slowapi==0.1.9
psycopg2-binary==2.9.10
bleach==6.1.0
